w3d2_utils.py
import random
import numpy as np
import torch as t
import gym


def set_seed(seed):
    random.seed(seed)
    np.random.seed(seed)
    t.manual_seed(seed)


def make_env(env_id: str, seed: int, idx: int, capture_video: bool, run_name: str):
    """Return a function that creates the specified environment inside convenience wrappers."""

    def thunk():
        env = gym.make(env_id)
        env = gym.wrappers.RecordEpisodeStatistics(env)
        if capture_video:
            if idx == 0:
                env = gym.wrappers.RecordVideo(env, f"videos/{run_name}")
        env.reset(seed=seed)
        env.action_space.seed(seed)
        env.observation_space.seed(seed)
        return env

    return thunk
